+++ src/app/app.module.ts
@@ -13,7 +13,7 @@ import { TestCaseEffect } from './Domain/state/test-case/test-case.effect';
 import { UserComponent } from './UI/user/user.component';
 import { UserEffect } from './Domain/state/user/user.effect';
 import { MainComponent } from './UI/main/main.component';
-import { ReactiveFormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 @NgModule({
   declarations: [
@@ -29,6 +29,7 @@ import { ReactiveFormsModule } from '@angular/forms';
     EffectsModule.forRoot([ForecastDataEffect, UserEffect, TestCaseEffect]),
     StoreDevtoolsModule.instrument({ maxAge: 25, logOnly: !isDevMode() }),
     ReactiveFormsModule,
+    FormsModule,
   ],
   providers: [],
   bootstrap: [AppComponent]